CH4/OFI: Reduce immediate data field to 32 bits
The OFI spec only requires providers to supply 4 bytes for
`cq_data_size`. We allocate 64 bits for the dest argument despite the
fact that we never use more than 32 bits.

Change this to use a uint32_t for the dest argument so we don't overrun
the provider's CQ data field.

Fixes pmodels/mpich#2478

+++ src/mpid/ch4/netmod/ofi/ofi_impl.h
@@ -405,7 +405,7 @@ MPL_STATIC_INLINE_PREFIX MPIR_Request *MPIDI_OFI_context_to_request(void *contex
 #undef FCNAME
 #define FCNAME MPL_QUOTE(FUNCNAME)
 MPL_STATIC_INLINE_PREFIX int MPIDI_OFI_send_handler(struct fid_ep *ep, const void *buf, size_t len,
-                                             void *desc, uint64_t dest, fi_addr_t dest_addr,
+                                             void *desc, uint32_t dest, fi_addr_t dest_addr,
                                              uint64_t tag, void *context, int is_inject,
                                              int do_lock)
 {
